requirements.txt
tencentcloud-sdk-python>=3.0.0
flask>=3.0.0
flask-cors>=6.0.0

